@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <RootNamespace></RootNamespace>
     <NoWarn>CS0649;CS0169</NoWarn>
     <NukeRootDirectory>..</NukeRootDirectory>
@@ -11,16 +11,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Nuke.Common" Version="6.2.1" />
+    <PackageReference Include="Nuke.Common" Version="6.3.0" PrivateAssets="all" />
+    <PackageReference Include="Candoumbe.Pipelines" Version="0.3.0-beta0001" />
     <PackageReference Include="ReportGenerator" Version="5.1.9" />
     <PackageDownload Include="GitVersion.Tool" Version="[5.8.0]" />
     <PackageDownload Include="Codecov.Tool" Version="[1.12.3]" />
-    <PackageDownload Include="dotnet-stryker" Version="[3.0.1]"/>
-    
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\src\Candoumbe.MiscUtilities\Candoumbe.MiscUtilities.csproj" />
+    <PackageDownload Include="dotnet-stryker" Version="[3.0.1]" />
   </ItemGroup>
 
 </Project>
