@@ -1,4 +1,7 @@
-﻿using System.Collections;
+﻿// "Copyright (c) Cyrille NDOUMBE.
+// Licenced under GNU General Public Licence, version 3.0"
+
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
@@ -167,7 +170,7 @@ namespace System
         /// <typeparam name="TDest">targeted type</typeparam>
         /// <param name="obj">The object to cast</param>
         /// <returns>The "safe cast" result</returns>
-        /// <exception cref="ArgumentNullException">if <paramref name="obj"/> is <c>null</c></exception>
+        /// <exception cref="ArgumentNullException">if <paramref name="obj"/> is <see langword="null"/></exception>
         public static TDest As<TDest>(this object obj) => (TDest)As(obj, typeof(TDest));
 
         /// <summary>
@@ -176,7 +179,7 @@ namespace System
         /// <param name="obj">The object to cast</param>
         /// <param name="targetType">type to cast </param>
         /// <returns>The "safe cast" result</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="targetType"/> is <c>null</c></exception>
+        /// <exception cref="ArgumentNullException"><paramref name="targetType"/> is <see langword="null"/></exception>
         public static object As(this object obj, Type targetType)
         {
             if (targetType == null)
@@ -216,6 +219,9 @@ namespace System
         public static string Jsonify(this object obj, JsonSerializerOptions settings = null)
         {
             string json = null;
+#if NET7_0_OR_GREATER
+            settings ??= JsonSerializerOptions.Default;
+#endif
             if (obj is not null)
             {
 #if REQUIRES_DATE_AND_TIME_ONLY_SERIALIZATION_WORKAROUND
@@ -244,5 +250,5 @@ namespace System
             return json;
         }
 #endif
-            }
+    }
 }
