@@ -9,8 +9,8 @@ namespace Candoumbe.MiscUtilities.PerformanceTests;
 [MemoryDiagnoser]
 [RPlotExporter]
 [SimpleJob(RuntimeMoniker.NetCoreApp31)]
-[SimpleJob(RuntimeMoniker.Net50)]
 [SimpleJob(RuntimeMoniker.Net60)]
+[SimpleJob(RuntimeMoniker.Net70)]
 public class StringExtensionsTests
 {
 
